+++ src/test/java/com/google/cloud/tools/gradle/appengine/AppEnginePluginTest.java
@@ -28,6 +28,7 @@ import com.google.cloud.tools.gradle.appengine.standard.AppEngineStandardPlugin;
 import java.io.IOException;
 import org.gradle.api.Project;
 import org.gradle.testkit.runner.BuildResult;
+import org.gradle.testkit.runner.UnexpectedBuildFailure;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -36,6 +37,29 @@ import org.junit.rules.TemporaryFolder;
 public class AppEnginePluginTest {
 
   @Rule public TemporaryFolder testProjectRoot = new TemporaryFolder();
+
+  @Test
+  public void testCheckGradleVersion_pass() {
+    new TestProject(testProjectRoot.getRoot())
+        .applyGradleRunnerWithVersion(AppEnginePlugin.GRADLE_MIN_VERSION.getVersion());
+    // pass
+  }
+
+  @Test
+  public void testCheckGradleVersion_fail() throws IOException {
+    try {
+      new TestProject(testProjectRoot.getRoot())
+          .addAutoDetectingBuildFile()
+          .applyGradleRunnerWithVersion("2.8");
+    } catch (UnexpectedBuildFailure ex) {
+      assertThat(
+          ex.getMessage(),
+          containsString(
+              "Detected Gradle 2.8, but the appengine-gradle-plugin requires "
+                  + AppEnginePlugin.GRADLE_MIN_VERSION
+                  + " or higher."));
+    }
+  }
 
   @Test
   public void testDetectStandard_withGradleRunner() throws IOException {